completed review, contact, and footer sections
review section will consist of 3 reviews from 3 different people. For a place holder I have placed myself for now with lorem text. The contact us page consists of a google map of the area the business is located as well as a name phone and email section for customers to contact the owner.
lastly the footer has a link to the instagram.