Added responsiveness to different sized resolutions
The final core step was completing the responsiveness for different resolutions. So when the screen gets smaller the header disappears and can be accessed in the menu icon. and did minor adjustments to the rest of the sections. I have shown the website at its current state to the bakery owner. Discussed minor preference changes and following steps would be to make an active cart system and fill out the Lorems.